refactor dashboard and slice models
