+++ superset/examples/helpers.py
@@ -25,13 +25,14 @@ from urllib import request
 from superset import app, db
 from superset.connectors.connector_registry import ConnectorRegistry
 from superset.models import core as models
+from superset.models.dashboard import Dashboard
+from superset.models.slice import Slice
 
 BASE_URL = "https://github.com/apache-superset/examples-data/blob/master/"
 
 # Shortcuts
 DB = models.Database
-Slice = models.Slice
-Dash = models.Dashboard
+Dash = Dashboard
 
 TBL = ConnectorRegistry.sources["table"]
