+++ superset/models/__init__.py
@@ -14,4 +14,4 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from . import core, schedules, sql_lab, user_attributes
+from . import core, schedules, slice, sql_lab, user_attributes
