+++ tests/dashboard_tests.py
@@ -25,6 +25,8 @@ from sqlalchemy import func
 from superset import db, security_manager
 from superset.connectors.sqla.models import SqlaTable
 from superset.models import core as models
+from superset.models.dashboard import Dashboard
+from superset.models.slice import Slice
 
 from .base_tests import SupersetTestCase
 
@@ -59,23 +61,23 @@ class DashboardTests(SupersetTestCase):
     def test_dashboard(self):
         self.login(username="admin")
         urls = {}
-        for dash in db.session.query(models.Dashboard).all():
+        for dash in db.session.query(Dashboard).all():
             urls[dash.dashboard_title] = dash.url
         for title, url in urls.items():
             assert escape(title) in self.client.get(url).data.decode("utf-8")
 
     def test_new_dashboard(self):
         self.login(username="admin")
-        dash_count_before = db.session.query(func.count(models.Dashboard.id)).first()[0]
+        dash_count_before = db.session.query(func.count(Dashboard.id)).first()[0]
         url = "/dashboard/new/"
         resp = self.get_resp(url)
         self.assertIn("[ untitled dashboard ]", resp)
-        dash_count_after = db.session.query(func.count(models.Dashboard.id)).first()[0]
+        dash_count_after = db.session.query(func.count(Dashboard.id)).first()[0]
         self.assertEqual(dash_count_before + 1, dash_count_after)
 
     def test_dashboard_modes(self):
         self.login(username="admin")
-        dash = db.session.query(models.Dashboard).filter_by(slug="births").first()
+        dash = db.session.query(Dashboard).filter_by(slug="births").first()
         url = dash.url
         if dash.url.find("?") == -1:
             url += "?"
@@ -88,7 +90,7 @@ class DashboardTests(SupersetTestCase):
 
     def test_save_dash(self, username="admin"):
         self.login(username=username)
-        dash = db.session.query(models.Dashboard).filter_by(slug="births").first()
+        dash = db.session.query(Dashboard).filter_by(slug="births").first()
         positions = self.get_mock_positions(dash)
         data = {
             "css": "",
@@ -102,7 +104,7 @@ class DashboardTests(SupersetTestCase):
 
     def test_save_dash_with_filter(self, username="admin"):
         self.login(username=username)
-        dash = db.session.query(models.Dashboard).filter_by(slug="world_health").first()
+        dash = db.session.query(Dashboard).filter_by(slug="world_health").first()
 
         positions = self.get_mock_positions(dash)
         filters = {str(dash.slices[0].id): {"region": ["North America"]}}
@@ -119,9 +121,7 @@ class DashboardTests(SupersetTestCase):
         resp = self.get_resp(url, data=dict(data=json.dumps(data)))
         self.assertIn("SUCCESS", resp)
 
-        updatedDash = (
-            db.session.query(models.Dashboard).filter_by(slug="world_health").first()
-        )
+        updatedDash = db.session.query(Dashboard).filter_by(slug="world_health").first()
         new_url = updatedDash.url
         self.assertIn("region", new_url)
 
@@ -130,7 +130,7 @@ class DashboardTests(SupersetTestCase):
 
     def test_save_dash_with_invalid_filters(self, username="admin"):
         self.login(username=username)
-        dash = db.session.query(models.Dashboard).filter_by(slug="world_health").first()
+        dash = db.session.query(Dashboard).filter_by(slug="world_health").first()
 
         # add an invalid filter slice
         positions = self.get_mock_positions(dash)
@@ -148,15 +148,13 @@ class DashboardTests(SupersetTestCase):
         resp = self.get_resp(url, data=dict(data=json.dumps(data)))
         self.assertIn("SUCCESS", resp)
 
-        updatedDash = (
-            db.session.query(models.Dashboard).filter_by(slug="world_health").first()
-        )
+        updatedDash = db.session.query(Dashboard).filter_by(slug="world_health").first()
         new_url = updatedDash.url
         self.assertNotIn("region", new_url)
 
     def test_save_dash_with_dashboard_title(self, username="admin"):
         self.login(username=username)
-        dash = db.session.query(models.Dashboard).filter_by(slug="births").first()
+        dash = db.session.query(Dashboard).filter_by(slug="births").first()
         origin_title = dash.dashboard_title
         positions = self.get_mock_positions(dash)
         data = {
@@ -167,9 +165,7 @@ class DashboardTests(SupersetTestCase):
         }
         url = "/superset/save_dash/{}/".format(dash.id)
         self.get_resp(url, data=dict(data=json.dumps(data)))
-        updatedDash = (
-            db.session.query(models.Dashboard).filter_by(slug="births").first()
-        )
+        updatedDash = db.session.query(Dashboard).filter_by(slug="births").first()
         self.assertEqual(updatedDash.dashboard_title, "new title")
         # bring back dashboard original title
         data["dashboard_title"] = origin_title
@@ -177,7 +173,7 @@ class DashboardTests(SupersetTestCase):
 
     def test_save_dash_with_colors(self, username="admin"):
         self.login(username=username)
-        dash = db.session.query(models.Dashboard).filter_by(slug="births").first()
+        dash = db.session.query(Dashboard).filter_by(slug="births").first()
         positions = self.get_mock_positions(dash)
         new_label_colors = {"data value": "random color"}
         data = {
@@ -191,9 +187,7 @@ class DashboardTests(SupersetTestCase):
         }
         url = "/superset/save_dash/{}/".format(dash.id)
         self.get_resp(url, data=dict(data=json.dumps(data)))
-        updatedDash = (
-            db.session.query(models.Dashboard).filter_by(slug="births").first()
-        )
+        updatedDash = db.session.query(Dashboard).filter_by(slug="births").first()
         self.assertIn("color_namespace", updatedDash.json_metadata)
         self.assertIn("color_scheme", updatedDash.json_metadata)
         self.assertIn("label_colors", updatedDash.json_metadata)
@@ -205,7 +199,7 @@ class DashboardTests(SupersetTestCase):
 
     def test_copy_dash(self, username="admin"):
         self.login(username=username)
-        dash = db.session.query(models.Dashboard).filter_by(slug="births").first()
+        dash = db.session.query(Dashboard).filter_by(slug="births").first()
         positions = self.get_mock_positions(dash)
         new_label_colors = {"data value": "random color"}
         data = {
@@ -223,7 +217,7 @@ class DashboardTests(SupersetTestCase):
         dash_id = dash.id
         url = "/superset/save_dash/{}/".format(dash_id)
         self.client.post(url, data=dict(data=json.dumps(data)))
-        dash = db.session.query(models.Dashboard).filter_by(id=dash_id).first()
+        dash = db.session.query(Dashboard).filter_by(id=dash_id).first()
         orig_json_data = dash.data
 
         # Verify that copy matches original
@@ -241,16 +235,12 @@ class DashboardTests(SupersetTestCase):
 
     def test_add_slices(self, username="admin"):
         self.login(username=username)
-        dash = db.session.query(models.Dashboard).filter_by(slug="births").first()
+        dash = db.session.query(Dashboard).filter_by(slug="births").first()
         new_slice = (
-            db.session.query(models.Slice)
-            .filter_by(slice_name="Energy Force Layout")
-            .first()
+            db.session.query(Slice).filter_by(slice_name="Energy Force Layout").first()
         )
         existing_slice = (
-            db.session.query(models.Slice)
-            .filter_by(slice_name="Girl Name Cloud")
-            .first()
+            db.session.query(Slice).filter_by(slice_name="Girl Name Cloud").first()
         )
         data = {
             "slice_ids": [new_slice.data["slice_id"], existing_slice.data["slice_id"]]
@@ -259,23 +249,21 @@ class DashboardTests(SupersetTestCase):
         resp = self.client.post(url, data=dict(data=json.dumps(data)))
         assert "SLICES ADDED" in resp.data.decode("utf-8")
 
-        dash = db.session.query(models.Dashboard).filter_by(slug="births").first()
+        dash = db.session.query(Dashboard).filter_by(slug="births").first()
         new_slice = (
-            db.session.query(models.Slice)
-            .filter_by(slice_name="Energy Force Layout")
-            .first()
+            db.session.query(Slice).filter_by(slice_name="Energy Force Layout").first()
         )
         assert new_slice in dash.slices
         assert len(set(dash.slices)) == len(dash.slices)
 
         # cleaning up
-        dash = db.session.query(models.Dashboard).filter_by(slug="births").first()
+        dash = db.session.query(Dashboard).filter_by(slug="births").first()
         dash.slices = [o for o in dash.slices if o.slice_name != "Energy Force Layout"]
         db.session.commit()
 
     def test_remove_slices(self, username="admin"):
         self.login(username=username)
-        dash = db.session.query(models.Dashboard).filter_by(slug="births").first()
+        dash = db.session.query(Dashboard).filter_by(slug="births").first()
         origin_slices_length = len(dash.slices)
 
         positions = self.get_mock_positions(dash)
@@ -297,7 +285,7 @@ class DashboardTests(SupersetTestCase):
         dash_id = dash.id
         url = "/superset/save_dash/{}/".format(dash_id)
         self.client.post(url, data=dict(data=json.dumps(data)))
-        dash = db.session.query(models.Dashboard).filter_by(id=dash_id).first()
+        dash = db.session.query(Dashboard).filter_by(id=dash_id).first()
 
         # verify slices data
         data = dash.data
@@ -307,7 +295,7 @@ class DashboardTests(SupersetTestCase):
         table = db.session.query(SqlaTable).filter_by(table_name="birth_names").one()
 
         # Make the births dash published so it can be seen
-        births_dash = db.session.query(models.Dashboard).filter_by(slug="births").one()
+        births_dash = db.session.query(Dashboard).filter_by(slug="births").one()
         births_dash.published = True
 
         db.session.merge(births_dash)
@@ -345,7 +333,7 @@ class DashboardTests(SupersetTestCase):
         table = db.session.query(SqlaTable).filter_by(table_name="birth_names").one()
         self.grant_public_access_to_table(table)
 
-        dash = db.session.query(models.Dashboard).filter_by(slug="births").first()
+        dash = db.session.query(Dashboard).filter_by(slug="births").first()
         dash.owners = [security_manager.find_user("admin")]
         dash.created_by = security_manager.find_user("admin")
         db.session.merge(dash)
@@ -354,7 +342,7 @@ class DashboardTests(SupersetTestCase):
         assert "Births" in self.get_resp("/superset/dashboard/births/")
 
     def test_only_owners_can_save(self):
-        dash = db.session.query(models.Dashboard).filter_by(slug="births").first()
+        dash = db.session.query(Dashboard).filter_by(slug="births").first()
         dash.owners = []
         db.session.merge(dash)
         db.session.commit()
@@ -365,18 +353,16 @@ class DashboardTests(SupersetTestCase):
 
         alpha = security_manager.find_user("alpha")
 
-        dash = db.session.query(models.Dashboard).filter_by(slug="births").first()
+        dash = db.session.query(Dashboard).filter_by(slug="births").first()
         dash.owners = [alpha]
         db.session.merge(dash)
         db.session.commit()
         self.test_save_dash("alpha")
 
     def test_owners_can_view_empty_dashboard(self):
-        dash = (
-            db.session.query(models.Dashboard).filter_by(slug="empty_dashboard").first()
-        )
+        dash = db.session.query(Dashboard).filter_by(slug="empty_dashboard").first()
         if not dash:
-            dash = models.Dashboard()
+            dash = Dashboard()
             dash.dashboard_title = "Empty Dashboard"
             dash.slug = "empty_dashboard"
         else:
@@ -394,9 +380,7 @@ class DashboardTests(SupersetTestCase):
     def test_users_can_view_published_dashboard(self):
         table = db.session.query(SqlaTable).filter_by(table_name="energy_usage").one()
         # get a slice from the allowed table
-        slice = (
-            db.session.query(models.Slice).filter_by(slice_name="Energy Sankey").one()
-        )
+        slice = db.session.query(Slice).filter_by(slice_name="Energy Sankey").one()
 
         self.grant_public_access_to_table(table)
 
@@ -404,13 +388,13 @@ class DashboardTests(SupersetTestCase):
         published_dash_slug = f"published_dash_{random()}"
 
         # Create a published and hidden dashboard and add them to the database
-        published_dash = models.Dashboard()
+        published_dash = Dashboard()
         published_dash.dashboard_title = "Published Dashboard"
         published_dash.slug = published_dash_slug
         published_dash.slices = [slice]
         published_dash.published = True
 
-        hidden_dash = models.Dashboard()
+        hidden_dash = Dashboard()
         hidden_dash.dashboard_title = "Hidden Dashboard"
         hidden_dash.slug = hidden_dash_slug
         hidden_dash.slices = [slice]
@@ -430,13 +414,13 @@ class DashboardTests(SupersetTestCase):
         not_my_dash_slug = f"not_my_dash_{random()}"
 
         # Create one dashboard I own and another that I don't
-        dash = models.Dashboard()
+        dash = Dashboard()
         dash.dashboard_title = "My Dashboard"
         dash.slug = my_dash_slug
         dash.owners = [user]
         dash.slices = []
 
-        hidden_dash = models.Dashboard()
+        hidden_dash = Dashboard()
         hidden_dash.dashboard_title = "Not My Dashboard"
         hidden_dash.slug = not_my_dash_slug
         hidden_dash.slices = []
@@ -457,11 +441,11 @@ class DashboardTests(SupersetTestCase):
         fav_dash_slug = f"my_favorite_dash_{random()}"
         regular_dash_slug = f"regular_dash_{random()}"
 
-        favorite_dash = models.Dashboard()
+        favorite_dash = Dashboard()
         favorite_dash.dashboard_title = "My Favorite Dashboard"
         favorite_dash.slug = fav_dash_slug
 
-        regular_dash = models.Dashboard()
+        regular_dash = Dashboard()
         regular_dash.dashboard_title = "A Plain Ol Dashboard"
         regular_dash.slug = regular_dash_slug
 
@@ -469,7 +453,7 @@ class DashboardTests(SupersetTestCase):
         db.session.merge(regular_dash)
         db.session.commit()
 
-        dash = db.session.query(models.Dashboard).filter_by(slug=fav_dash_slug).first()
+        dash = db.session.query(Dashboard).filter_by(slug=fav_dash_slug).first()
 
         favorites = models.FavStar()
         favorites.obj_id = dash.id
@@ -490,7 +474,7 @@ class DashboardTests(SupersetTestCase):
         slug = f"admin_owned_unpublished_dash_{random()}"
 
         # Create a dashboard owned by admin and unpublished
-        dash = models.Dashboard()
+        dash = Dashboard()
         dash.dashboard_title = "My Dashboard"
         dash.slug = slug
         dash.owners = [admin_user]
