+++ tests/schedules_test.py
@@ -23,7 +23,7 @@ from selenium.common.exceptions import WebDriverException
 
 from tests.test_app import app
 from superset import db
-from superset.models.core import Dashboard, Slice
+from superset.models.dashboard import Dashboard
 from superset.models.schedules import (
     DashboardEmailSchedule,
     EmailDeliveryType,
@@ -36,6 +36,7 @@ from superset.tasks.schedules import (
     deliver_slice,
     next_schedules,
 )
+from superset.models.slice import Slice
 from tests.base_tests import SupersetTestCase
 
 from .utils import read_fixture
